Refactor the Front-end to use 'routes' :art:
On the previous structure the code wasn't using well the React.js.
All the html was on the index.html.

Now the index.html just have some main scripts that we're going to
use on our Front-end - such as Bootstrap.css - and its just renders
the 'app' id.

The app id calls the App.js that divides the containers on routes.
On the first scenario we're going to have three routes: /, /login and
/chatbot that renders the containers <Login /> and <ChatbotDash />
respectively.

The ChatbotDash calls the Chatbot screen that uses the Chatbot service.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>