Add DBClient service on the Back-End :heavy_plus_sign:
This class receives a db name as it parameter or by opts.

The DBClient.insert method receives a document to insert
on the db set on the constructor or in its options. If the
document is a valid object, it is registered on the db,
otherwise the method throws an error.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>