Add EmailClient service on the Back-End :heavy_plus_sign:
This class uses the 'nodemailer' lib to send email.

The smtp server choosed was the GMAIL. Basically on the
constructor the user can specifye a 'from' email and a
'to' email - both string type. If the user doesn't specifye
anything, is assigned the default email on the environment
variables.

The method EmailClient.send receives a subjectMail and a
bodyMail, then if both are type string this method assigned
both paremeters on the EmailClient.options and sends an email
to the 'options.to' parameter. Otherwise it will throw an error.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>