Closes #14 - Add BOX service on the Back-End :heavy_plus_sign:
This class receives environment variables to configure the connection
to our Box instance on the web.

The getPublicLink is responsible to generate embed links to the documents
on our Box Storage, allowing to render a pdf file on the answer of the
chatbot. The id passed to it must be a number, otherwise it throws an error.

To configure a Box API you can see this [documentation](https://github.com/mentoring-projects/TOP-The-OCR-Project/wiki/Box-API-Step-by-Step)

Signed-off-by: Gustavo Porto Guedes<gustavo.guedes@fatec.sp.gov.br>