Closes #13 - Add search method to the DBClient service on the Back-End :heavy_plus_sign:
This method uses the Mongo db strategy(https://docs.mongodb.com/manual/tutorial/query-documents/)
to query documents in our db.
It must receive a valid object - with selector object key - to search the document on the Cloudant.
If it does not find anything, returns an empty array. In the context of the chatbot the object sent
to the Conversation service will be handled to make a query on our db, if finds something, it
changes the chatbot answer to what was designed on the db, otherwise it keeps the Conversation object
answer.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>