Closes #12 - Refactor the Conversation pipeline :heavy_plus_sign
Now when an input from the client side arrives to the Conversation Pipeline
it calls the conversation service, then with the payload the pipeline
extracts its intent and entities to search on the Cloudant db. If the Cloudant
db finds something the pipeline changes the answer and its output.cardType
to enriched. The BOX API is used to generate an embed link to the file id
stored on the db.

Signed-off-by: Gustavo Porto Guedes <gustavo.guedes@fatec.sp.gov.br>