Closes #15 - Create the feedback function in the dialog ending :heavy_plus_sign:
When the output.cardType from the Back-End is defined and is enriched,
the feedback element is rendered on the final of the chatbot's answer.
The user must click on the thumbs-up or the thumbs-down button, otherwise
the chatbot will not allow the user input. When the user clicks in one of
the buttons, its send a feedback by the Feedback service and if everything
went well sends a message to the Conversation service to continue the
conversation.